Move pkg/backends/oauth2/state to pkg/cache/state
