+++ pkg/backends/oauth2/configure.go
@@ -24,7 +24,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/greenpau/caddy-auth-portal/pkg/backends/oauth2/state"
+	"github.com/greenpau/caddy-auth-portal/pkg/cache/state"
 
 	"github.com/greenpau/caddy-auth-portal/pkg/errors"
 	"go.uber.org/zap"
