+++ pkg/cache/state/interface.go
