+++ pkg/cache/state/memory.go
